R Code
Please note that the R code only runs with the complete dataset (especially for Figures 1 and 3, and the Supplementary Tables). The full dataset can be provided confidentially upon request.

If you use the public version of the data, you will need to adapt the code for Figures 1 and 3, as well as Supplementary Tables 1 and 2 (by removing the human data). Consequently, ANOVA and correlation results may differ.

Data for Figure 2 are fully available without restriction. Data for Figure 4 are included, but the plot was not generated in R.